Closest Points (divide and conquer)
Problem Introduction
In this problem, your goal is to find the closest pair of points among the given 𝑛 points. This is a basic primitive in computational geometry having applications in, for example, graphics, computer vision, traffic-control systems.

Problem Description
Task. Given 𝑛 points on a plane, find the smallest distance between a pair of two (different) points. Recall that the distance between points (𝑥1, 𝑦1) and (𝑥2, 𝑦2) is equal to sqrt((𝑥1 − 𝑥2)^2 + (𝑦1 − 𝑦2)^2).

Input Format. The first line contains the number 𝑛 of points. Each of the following 𝑛 lines defines a point (𝑥𝑖, 𝑦𝑖).

Constraints. 2 ≤ 𝑛 ≤ 10^5 ; −10^9 ≤ 𝑥𝑖 , 𝑦𝑖 ≤ 10^9 are integers.

Output Format. Output the minimum distance. The absolute value of the difference between the answer of your program and the optimal value should be at most 10^(-3). To ensure this, output your answer with at least four digits after the decimal point (otherwise your answer, while being computed correctly, can turn out to be wrong because of rounding issues).

Sample 1.
Input:
2
0 0
3 4
Output:
5.0
There are only two points here. The distance between them is 5.

Sample 2.
Input:
4
7 7
1 1
0 0
4 8
7 7
Output:
0.0
There are two coinciding points among the four given points. Thus, the minimum distance is zero.

Sample 3.
Input:
11 
4 4 
-2 -2 
-3 -4 
-1 3 
2 3
-4 0 
1 1 
-1 -1
3 -1 
-4 2 
-2 4
Output:
1.414213
The smallest distance is  sqrt(2).
There are two pairs of points at this distance: (−1,−1) and (−2,−2); (−2, 4) and (−1, 3)

What to do
This computational geometry problem has many applications in computer graphics and vision. A naive algorithm with quadratic running time iterates through all pairs of points to find the closest pair. Your goal is to design an 𝑂(𝑛 log 𝑛) time divide and conquer algorithm.
To solve this problem in time 𝑂(𝑛 log 𝑛), let’s first split the given 𝑛 points by an appropriately chosen vertical line into two halves 𝑆1 and 𝑆2 of size 𝑛 (assume for simplicity that all 𝑥-coordinates of the input points are different). By making two recursive calls for the sets 𝑆1 and 𝑆2, we find the minimum distances 𝑑1 and 𝑑2 in these subsets. Let 𝑑 = min{𝑑1, 𝑑2}. It remains to check whether there exist points 𝑝1 ∈ 𝑆1 and 𝑝2 ∈ 𝑆2 such that the distance between them is smaller than 𝑑. We cannot afford to check all possible such pairs since there are 𝑛 · 𝑛 = Θ(𝑛^2) of them.
To check this faster, we first discard all points from 𝑆1 and 𝑆2 whose 𝑥-distance to the middle line is greater
than 𝑑. That is, we focus on the following strip
Stop and think: Why can we narrow the search to this strip? 
Now, let’s sort the points of the strip by their 𝑦-coordinates and denote the resulting sorted list by 𝑃 = [𝑝1, . . . , 𝑝𝑘].
It turns out that if |𝑖 − 𝑗| > 7, then the distance between points 𝑝𝑖 and 𝑝𝑗 is greater than 𝑑 for sure. 
This follows from the Exercise Break below.
Exercise break: Partition the strip into 𝑑 × 𝑑 squares as shown below and show that each such square contains at most four input points.
This results in the following algorithm. We first sort the given 𝑛 points by their 𝑥-coordinates and then split the resulting sorted list into two halves 𝑆1 and 𝑆2 of size 𝑛/2. By making a recursive call for each of the sets 𝑆1 and 𝑆2, we find the minimum distances 𝑑1 and 𝑑2 in them. Let 𝑑 = min{𝑑1 , 𝑑2}. However, we are not done yet as we also need to find the minimum distance between points from different sets (i.e, a point from 𝑆1 and a point from 𝑆2) and check whether it is smaller than 𝑑. To perform such a check, we filter the initial point set and keep only those points whose 𝑥-distance to the middle line does not exceed 𝑑. Afterwards, we sort the set of points in the resulting strip by their 𝑦-coordinates and scan the resulting list of points. For each point, we compute its distance to the seven subsequent points in this list and compute 𝑑′, the minimum distance that we encountered during this scan. Afterwards, we return min{𝑑,𝑑′}.
The running time of the algorithm satisfies the recurrence relation
𝑇 (𝑛) = 2 · 𝑇 (︁𝑛/2)︁ + 𝑂(𝑛 log 𝑛) . 
The 𝑂(𝑛 log 𝑛) term comes from sorting the points in the strip by their 𝑦-coordinates at every iteration. 
Exercise break: Prove that 𝑇 (𝑛) = 𝑂(𝑛 log^2 𝑛) by analyzing the recursion tree of the algorithm. 
Exercise break: Show how to bring the running time down to 𝑂(𝑛log𝑛) by avoiding sorting at each recursive call. (can be done by storing points sorted by y-coordinate separately and merging them)